Add files via upload
simple demonstration of using GPT4all with a modal.com instance, it retrieves the 'live' json list of models, makes that available as a drop down, a single file app.py handles all the code for interacting with modal and uses flask to create endpoints to ask a question, which sets the model, you can switch model at any time, however this will lose the current context, if you wish to continue the conversation do not change the model, or copy and paste the conversation so far and send that to the new model. There is an index.html file that handles the simple UI and uses javascript to handle the data submission and display the results.